requirements.txt
discord
asyncpg
python-dotenv
requests
yagmail
asyncio